requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.3
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.5.0
passlib==1.7.4
psycopg==3.2.11
psycopg-binary==3.2.11
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
redis==5.2.0
rq==1.16.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
pytest==7.4.0
